rename file names using '-' and lowcase
