+++ client/src/app.ts
@@ -1,5 +1,5 @@
 import './styles/common.scss';
-import TodoList from './pages/todoList';
+import TodoList from './pages/todo-list';
 import $modal from './components/modal';
 
 // test
@@ -7,18 +7,17 @@ type pageType = 'todoList' | 'login';
 const page: pageType = 'todoList';
 
 class App {
-  private container!: HTMLElement;
+	private container!: HTMLElement;
 
-  constructor(container: HTMLElement) {
-    this.container = container;
-    this.container.appendChild($modal);
-    if (page === 'todoList') {
-      const todoList = new TodoList();
-      console.log(todoList);
-      this.container.appendChild(todoList);
-    }
-  }
+	constructor(container: HTMLElement) {
+		this.container = container;
+		this.container.appendChild($modal);
+		if (page === 'todoList') {
+			const todoList = new TodoList();
+			console.log(todoList);
+			this.container.appendChild(todoList);
+		}
+	}
 }
 
 new App(document.querySelector('#app') as HTMLElement);
-
