+++ client/src/components/card-input/index.ts
@@ -1,0 +1,4 @@
+import './card-input.scss';
+import CardInput from './card-input';
+
+export default CardInput;
