+++ client/src/components/card/card.ts
@@ -1,5 +1,16 @@
 import $modal from '../modal';
-import CardInterface from './card_interface';
+
+export interface CardInterface {
+	card_id: number;
+	order_weight: number;
+	title: string;
+	user_name: string;
+	content: string;
+	last_update: number;
+	create_date: number;
+	user_id: number;
+	topic_id: number;
+}
 
 class Card extends HTMLElement {
 	private state: CardInterface;
