--- client/src/components/card/card_interface.ts
@@ -1,11 +1,0 @@
-export default interface CardInterface {
-	card_id: number;
-	order_weight: number;
-	title: string;
-	user_name: string;
-	content: string;
-	last_update: number;
-	create_date: number;
-	user_id: number;
-	topic_id: number;
-}
