--- client/src/components/card_input/index.ts
@@ -1,4 +1,0 @@
-import './card_input.scss';
-import CardInput from './card_input';
-
-export default CardInput;
