--- client/src/components/modal/modal_interface.ts
@@ -1,6 +1,0 @@
-export default interface ModalInterface {
-	title: string;
-	content: string;
-	resolve?: string;
-	reject?: string;
-}
