--- client/src/components/sidebar/activity_interface.ts
@@ -1,12 +1,0 @@
-export default interface ActivityInterface {
-	activity_id: number;
-	service_id: number;
-	card_id: number;
-	user_id: number;
-	uid: string;
-	action: string;
-	card_title: string;
-	from: string;
-	to: string;
-	create_date: number;
-}
