--- client/src/components/topic/topic_interface.ts
@@ -1,6 +1,0 @@
-export default interface TopicInterface {
-	topic_id: number;
-	order_weight: number;
-	title: string;
-	count: number;
-}
