+++ client/src/pages/todo-list/index.ts
@@ -1,0 +1,4 @@
+import './todo-list.scss';
+import TodoList from './todo-list';
+
+export default TodoList;
