--- client/src/pages/todoList/index.ts
@@ -1,4 +1,0 @@
-import './todoList.scss';
-import TodoList from './todoList';
-
-export default TodoList;
